Store deletion and modified date for series in the DB
This is useful for several things, but in particular: we need this for
the upcoming Tobira harvesting API. Without information about when
something about a series changes, a syncing application would have to
request all series every time, which is quite a waste.

To also be able to notify other applications about deleted series, we
cannot delete them for real, but just flag them as deleted. Instead of
a flag, a timestamp is stored as that might be useful in the future.