Remove XML -> string -> XML roundtrip from `MediaPackageImpl.valueOf`
This was introduced in this commit:
https://github.com/opencast/opencast/commit/2c6a0bcadeceb430e9d2a7c06f2a108844f4727e#diff-2dc988e723ae8980e3192e56d0f8ff2a072dacb716f9145ad3737f47a3f722a3L1448

"Fixed namespace handling in almost all internal XML structures"

I asked in the community chat and it seems like we don't need that
anymore:

> From memory and based on this patch being introduced in for 1.4.1-rc6:
> I remember that there was an old version of media packages which had
> different namespaces or a different namespace structure and that we
> had some code for still being able to handle old media package XMLs.
>
> My guess would be that basically nobody still needs this because we
> talk about Matterhorn 1.3 media packages. Disabling this by default
> would probably be reasonable

So here we go. We do this for performance, of course. I actually found
this piece while tracking down why some operation was so terribly slow.
It was mostly due to this method. From my very basic measurements, this
change improves loading events from the search index by a factor of at
least 4x. Previously, loading 500 events took in the neighborhood of
1s.